+++ NEWS.md
@@ -1,4 +1,4 @@
-# 0.5.1.9000
+# 0.6.0
 
 * New function `startApp()` to start the app running on local machine